Replaced aspect_w and aspect_h with aspect_ratio in EDLSession
Asset already has asset_ratio.
Aspect width and height is calculated for gui

+++ cinelerra/pluginvclient.h
@@ -116,8 +116,8 @@ public:
 	int project_color_model;
 
 // Aspect ratio
-	float aspect_w;
-	float aspect_h;
+	double aspect_w;
+	double aspect_h;
 
 // Tempo
 	VFrame *temp;
